feat: add beforeBuild and afterBuild in doc plugin
