+++ .changeset/light-dragons-occur.md
@@ -1,0 +1,6 @@
+---
+'@modern-js/doc-core': minor
+---
+
+feat: add beforeBuild and afterBuild in doc plugin
+feat: 在doc.plugin里面添加beforeBuild和afterBuild
